requirements.txt
uvicorn
fastapi
gitpython
openai
python-dotenv
response
jinja2
pinecone-client
numpy
fastapi_login
httpx
starlette
pydantic
boto3
scikit-learn
PyGithub
aiobotocore
chardet
